refactor: create WhispersAPI
