+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/hotf/HotfAPI.kt
@@ -2,20 +2,17 @@ package tech.thatgravyboat.skyblockapi.api.profile.hotf
 
 import me.owdding.ktmodules.Module
 import net.minecraft.world.item.Items
+import tech.thatgravyboat.skyblockapi.RemoveNextVersion
 import tech.thatgravyboat.skyblockapi.api.data.stored.HotfStorage
 import tech.thatgravyboat.skyblockapi.api.events.base.Subscription
-import tech.thatgravyboat.skyblockapi.api.events.base.predicates.OnlyWidget
-import tech.thatgravyboat.skyblockapi.api.events.info.TabWidget
-import tech.thatgravyboat.skyblockapi.api.events.info.TabWidgetChangeEvent
 import tech.thatgravyboat.skyblockapi.api.events.screen.InventoryChangeEvent
 import tech.thatgravyboat.skyblockapi.impl.tagkey.ItemModelTag
-import tech.thatgravyboat.skyblockapi.utils.extentions.*
+import tech.thatgravyboat.skyblockapi.utils.extentions.cleanName
+import tech.thatgravyboat.skyblockapi.utils.extentions.getItemModel
+import tech.thatgravyboat.skyblockapi.utils.extentions.getRawLore
+import tech.thatgravyboat.skyblockapi.utils.extentions.toIntValue
 import tech.thatgravyboat.skyblockapi.utils.regex.RegexGroup
 import tech.thatgravyboat.skyblockapi.utils.regex.RegexUtils.anyMatch
-import tech.thatgravyboat.skyblockapi.utils.regex.RegexUtils.match
-
-private const val MAIN_SLOT = 49
-private const val RESET_SLOT = 52
 
 @Module
 object HotfAPI {
@@ -43,22 +40,9 @@ object HotfAPI {
         "Heart of the Forest",
     )
 
-    private val whispersItemRegex = inventoryGroup.create(
-        "whispers.current",
-        "Forest Whispers: (?<amount>[\\d,.]+)",
-    )
-    private val whispersSpentItemRegex = inventoryGroup.create(
-        "whispers.spent",
-        "\\s*-\\s*(?<amount>[\\d,.]+) Forest Whispers",
-    )
-
     private val tokensRegex = inventoryGroup.create(
         "tokens",
         "Tokens of the Forest: (?<tokens>\\d+)",
-    )
-    private val forestWhispersRegex = RegexGroup.TABLIST_WIDGET.group("hotf").create(
-        "whispers",
-        "Forest Whispers: (?<amount>[\\w,.]+)",
     )
     //endregion
 
@@ -69,25 +53,16 @@ object HotfAPI {
     val activePerks: Map<String, HotfPerk>
         get() = perks.filter { it.value.unlocked && !it.value.disabled }
 
+    @RemoveNextVersion(ReplaceWith("Use WhispersAPI instead"))
     val whispers: Long
         get() = HotfStorage.whispers
 
+    @RemoveNextVersion(ReplaceWith("Use WhispersAPI instead"))
     val whispersTotal: Long
         get() = HotfStorage.whispersTotal
 
     val tokens: Int
         get() = HotfStorage.tokens
-
-    @Subscription
-    @OnlyWidget(TabWidget.FOREST_WHISPERS)
-    fun onTabWidgetChange(event: TabWidgetChangeEvent) {
-        forestWhispersRegex.anyMatch(event.new, "amount") { (amount) ->
-            val amount = amount.parseFormattedLong()
-            val diff = amount - HotfStorage.whispers
-            HotfStorage.whispers = amount
-            if (diff > 0) HotfStorage.whispersTotal += diff
-        }
-    }
 
     @Subscription
     fun onInventoryChange(event: InventoryChangeEvent) {
@@ -103,39 +78,15 @@ object HotfAPI {
             return
         }
 
-        if (event.item in ItemModelTag.HOTF_PERK_ITEMS) {
-            val model = event.item.getItemModel()
+        if (event.item !in ItemModelTag.HOTF_PERK_ITEMS) return
+        val model = event.item.getItemModel()
 
-            var level = 1
-            levelRegex.anyMatch(lore, "level") { (_level) ->
-                level = _level.toIntValue()
-            }
-            val disabled = disabledRegex.anyMatch(lore)
-            val unlocked = model != Items.MANGROVE_ROOTS && model != Items.PALE_OAK_SAPLING && model != Items.PALE_OAK_BUTTON
-            HotfStorage.setPerk(cleanName, HotfPerk(level, unlocked, disabled))
-            return
+        var level = 1
+        levelRegex.anyMatch(lore, "level") { (_level) ->
+            level = _level.toIntValue()
         }
-
-        // Whispers and Total Whispers
-        val mainItem = event.itemStacks.getOrNull(MAIN_SLOT) ?: return
-        val resetItem = event.itemStacks.getOrNull(RESET_SLOT)
-
-        var whispersTotal = 0L
-
-        val mainLore = mainItem.getRawLore()
-        for (line in mainLore) {
-            whispersItemRegex.match(line, "amount") { (amount) ->
-                HotfStorage.whispers = amount.toLongValue()
-                whispersTotal += this.whispers
-            }
-        }
-
-        if (resetItem != null) {
-            whispersSpentItemRegex.anyMatch(resetItem.getRawLore(), "amount") { (amount) ->
-                whispersTotal += amount.toLongValue()
-            }
-        }
-
-        HotfStorage.whispersTotal = whispersTotal
+        val disabled = disabledRegex.anyMatch(lore)
+        val unlocked = model != Items.MANGROVE_ROOTS && model != Items.PALE_OAK_SAPLING && model != Items.PALE_OAK_BUTTON
+        HotfStorage.setPerk(cleanName, HotfPerk(level, unlocked, disabled))
     }
 }
